pyproject.toml
[project]
name = "landseer_pipeline"
version = "0.1.0"
description = ""
authors = [
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "h5py (>=3.13.0,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "docker (<=7.1.0)",
    "torchattacks (>=3.5.1,<4.0.0)",
]

[tool.poetry]
packages = [{include = "*", from = "src"}]

[tool.poetry.scripts]
landseer = "main:main"

[tool.poetry.dependencies]
torch = {source = "torch"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = 'explicit'